+++ deployments/ropsten/CurvePool-LUSD-3CRV.json
@@ -1,5 +1,5 @@
 {
-  "address": "0x44fE67AcD955B2A782fB3f1957b56b80b492030F",
+  "address": "0x9F222Da5b7F1D7bD2987aB078A44e153eeAae66B",
   "abi": [
     {
       "anonymous": false,
@@ -209,34 +209,34 @@
       "type": "function"
     }
   ],
-  "transactionHash": "0x1dbe41cb9ed9cd1e0f41a16109f6318b25329dc1f2f971e32cbb51ab6afd063a",
+  "transactionHash": "0x9e1470a394631bd1af05dc9675e766b724ee3fad355f8b9f5c58d227ceed8880",
   "receipt": {
     "to": null,
     "from": "0x8f592F2594C43eA30879f9fbED4d930248D81D41",
-    "contractAddress": "0x44fE67AcD955B2A782fB3f1957b56b80b492030F",
-    "transactionIndex": 1,
+    "contractAddress": "0x9F222Da5b7F1D7bD2987aB078A44e153eeAae66B",
+    "transactionIndex": 3,
     "gasUsed": "716114",
-    "logsBloom": "0x00000000000000400000000000000000000000400000000000800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000001000000000000000000000000000000000000020000000000000000000800000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000000000000000000000000000200000000000000000000004000000000000000000000000000000000000000000000000000000000000000000000000000000000020000000000000000080000000000000000000000000000000000000000000000000",
-    "blockHash": "0x91037fb622115979605abf5634d08f24c7aa2f29ec7e7dde176cb25ab8319a85",
-    "transactionHash": "0x1dbe41cb9ed9cd1e0f41a16109f6318b25329dc1f2f971e32cbb51ab6afd063a",
+    "logsBloom": "0x00000000000000000000000000000000000000400000000000800000000000000000000000000000000000000000000000000000000000000000000000080000000000000000008000000000000000000001000000000000000000000000000000000000020000000000000000000800000000000000000000000000000000400000000000000400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000000000000000000000020000000000000000000002000000000000000000000000000000000000000000000",
+    "blockHash": "0x1ae0fc357131cbfb695246687e74d7a4e9f363cd45afa92e9fd9c01c09cc2bb5",
+    "transactionHash": "0x9e1470a394631bd1af05dc9675e766b724ee3fad355f8b9f5c58d227ceed8880",
     "logs": [
       {
-        "transactionIndex": 1,
-        "blockNumber": 12531878,
-        "transactionHash": "0x1dbe41cb9ed9cd1e0f41a16109f6318b25329dc1f2f971e32cbb51ab6afd063a",
-        "address": "0x44fE67AcD955B2A782fB3f1957b56b80b492030F",
+        "transactionIndex": 3,
+        "blockNumber": 12571141,
+        "transactionHash": "0x9e1470a394631bd1af05dc9675e766b724ee3fad355f8b9f5c58d227ceed8880",
+        "address": "0x9F222Da5b7F1D7bD2987aB078A44e153eeAae66B",
         "topics": [
           "0x8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0",
           "0x0000000000000000000000000000000000000000000000000000000000000000",
           "0x0000000000000000000000008f592f2594c43ea30879f9fbed4d930248d81d41"
         ],
         "data": "0x",
-        "logIndex": 2,
-        "blockHash": "0x91037fb622115979605abf5634d08f24c7aa2f29ec7e7dde176cb25ab8319a85"
+        "logIndex": 8,
+        "blockHash": "0x1ae0fc357131cbfb695246687e74d7a4e9f363cd45afa92e9fd9c01c09cc2bb5"
       }
     ],
-    "blockNumber": 12531878,
-    "cumulativeGasUsed": "823110",
+    "blockNumber": 12571141,
+    "cumulativeGasUsed": "1844858",
     "status": 1,
     "byzantium": true
   },
