+++ deployments/ropsten/YearnStrategy.json
@@ -1,5 +1,5 @@
 {
-  "address": "0x3ee890eEcB4277CAA05aba037973Cab13b3278DE",
+  "address": "0x0FBeFDD033371d540Fa4256a756D7A2604A0817a",
   "abi": [
     {
       "inputs": [
@@ -651,36 +651,36 @@
       "type": "function"
     }
   ],
-  "transactionHash": "0xd71d491c6776593e5c2b11da371d00799c6f4cb9aaa93a33b8d515dff5808407",
+  "transactionHash": "0x67e91a48aae52d090e58dbc3a94799f8a7ddc5c813ab1d44b2a57f448b46741d",
   "receipt": {
     "to": null,
     "from": "0x8f592F2594C43eA30879f9fbED4d930248D81D41",
-    "contractAddress": "0x3ee890eEcB4277CAA05aba037973Cab13b3278DE",
-    "transactionIndex": 6,
-    "gasUsed": "1446834",
-    "logsBloom": "0x00000004000000000000000000000000000000400000000000800400000000000000000000000000000000000000000000000000000000000000000000200000000000008020008000000000000000000001000000000000000000000000000000000000020000000004001000000900000000000200000000000000600000400000000000000000000000000010000000000000000000008000000000000000020000000000000200000000000000000020000004000000001000000000000000000000000000000000020000000000000000000000000100000000000020000110000400000000000000000000000000000000000000000000000000400200",
-    "blockHash": "0x85fc97e390003ea4d36c5f9662709d4d81d5d4347460b20f494acd3450dd4d95",
-    "transactionHash": "0xd71d491c6776593e5c2b11da371d00799c6f4cb9aaa93a33b8d515dff5808407",
+    "contractAddress": "0x0FBeFDD033371d540Fa4256a756D7A2604A0817a",
+    "transactionIndex": 15,
+    "gasUsed": "1446930",
+    "logsBloom": "0x00010004000000000004000000000000000000400000000000800000000000000000000000000000000000000000000000000000000000000000000000200000000000000020008200000000000002000001000400000000000000000000000000000000020000000004002000000800000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000100000020000000000000000000000000040000000000004000000001100000000000000000000000000400800020000000020000000000000000100000000000028000010000000100000000000000000000000000000000020000000000000000000",
+    "blockHash": "0xf5758601d9c942d6a26fb2a6647c7a1c192fc9ec9490d7f731ebde5f089c3506",
+    "transactionHash": "0x67e91a48aae52d090e58dbc3a94799f8a7ddc5c813ab1d44b2a57f448b46741d",
     "logs": [
       {
-        "transactionIndex": 6,
-        "blockNumber": 12531895,
-        "transactionHash": "0xd71d491c6776593e5c2b11da371d00799c6f4cb9aaa93a33b8d515dff5808407",
-        "address": "0x3ee890eEcB4277CAA05aba037973Cab13b3278DE",
+        "transactionIndex": 15,
+        "blockNumber": 12571159,
+        "transactionHash": "0x67e91a48aae52d090e58dbc3a94799f8a7ddc5c813ab1d44b2a57f448b46741d",
+        "address": "0x0FBeFDD033371d540Fa4256a756D7A2604A0817a",
         "topics": [
           "0x8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0",
           "0x0000000000000000000000000000000000000000000000000000000000000000",
           "0x0000000000000000000000008f592f2594c43ea30879f9fbed4d930248d81d41"
         ],
         "data": "0x",
-        "logIndex": 6,
-        "blockHash": "0x85fc97e390003ea4d36c5f9662709d4d81d5d4347460b20f494acd3450dd4d95"
+        "logIndex": 33,
+        "blockHash": "0xf5758601d9c942d6a26fb2a6647c7a1c192fc9ec9490d7f731ebde5f089c3506"
       },
       {
-        "transactionIndex": 6,
-        "blockNumber": 12531895,
-        "transactionHash": "0xd71d491c6776593e5c2b11da371d00799c6f4cb9aaa93a33b8d515dff5808407",
-        "address": "0x3ee890eEcB4277CAA05aba037973Cab13b3278DE",
+        "transactionIndex": 15,
+        "blockNumber": 12571159,
+        "transactionHash": "0x67e91a48aae52d090e58dbc3a94799f8a7ddc5c813ab1d44b2a57f448b46741d",
+        "address": "0x0FBeFDD033371d540Fa4256a756D7A2604A0817a",
         "topics": [
           "0x2f8788117e7eff1d82e926ec794901d17c78024a50270940304540a733656f0d",
           "0x0000000000000000000000000000000000000000000000000000000000000000",
@@ -688,49 +688,49 @@
           "0x0000000000000000000000008f592f2594c43ea30879f9fbed4d930248d81d41"
         ],
         "data": "0x",
-        "logIndex": 7,
-        "blockHash": "0x85fc97e390003ea4d36c5f9662709d4d81d5d4347460b20f494acd3450dd4d95"
+        "logIndex": 34,
+        "blockHash": "0xf5758601d9c942d6a26fb2a6647c7a1c192fc9ec9490d7f731ebde5f089c3506"
       },
       {
-        "transactionIndex": 6,
-        "blockNumber": 12531895,
-        "transactionHash": "0xd71d491c6776593e5c2b11da371d00799c6f4cb9aaa93a33b8d515dff5808407",
-        "address": "0x3ee890eEcB4277CAA05aba037973Cab13b3278DE",
+        "transactionIndex": 15,
+        "blockNumber": 12571159,
+        "transactionHash": "0x67e91a48aae52d090e58dbc3a94799f8a7ddc5c813ab1d44b2a57f448b46741d",
+        "address": "0x0FBeFDD033371d540Fa4256a756D7A2604A0817a",
         "topics": [
           "0x2f8788117e7eff1d82e926ec794901d17c78024a50270940304540a733656f0d",
           "0x241ecf16d79d0f8dbfb92cbc07fe17840425976cf0667f022fe9877caa831b08",
-          "0x00000000000000000000000060e8aead0b8e149633b240ea948ed50a81696750",
+          "0x000000000000000000000000a44567ea4b9649c86d18ae7edcb03d1752e792f7",
           "0x0000000000000000000000008f592f2594c43ea30879f9fbed4d930248d81d41"
         ],
         "data": "0x",
-        "logIndex": 8,
-        "blockHash": "0x85fc97e390003ea4d36c5f9662709d4d81d5d4347460b20f494acd3450dd4d95"
+        "logIndex": 35,
+        "blockHash": "0xf5758601d9c942d6a26fb2a6647c7a1c192fc9ec9490d7f731ebde5f089c3506"
       },
       {
-        "transactionIndex": 6,
-        "blockNumber": 12531895,
-        "transactionHash": "0xd71d491c6776593e5c2b11da371d00799c6f4cb9aaa93a33b8d515dff5808407",
-        "address": "0x99Fda92878c1d2f1e0971D1937C50CC578A33E3D",
+        "transactionIndex": 15,
+        "blockNumber": 12571159,
+        "transactionHash": "0x67e91a48aae52d090e58dbc3a94799f8a7ddc5c813ab1d44b2a57f448b46741d",
+        "address": "0x8D6C54d308C043Ee1914a4604d873E6538713AEF",
         "topics": [
           "0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925",
-          "0x0000000000000000000000003ee890eecb4277caa05aba037973cab13b3278de",
-          "0x0000000000000000000000003520a7e6d178b8f41bbe78766c6a939434bacf69"
+          "0x0000000000000000000000000fbefdd033371d540fa4256a756d7a2604a0817a",
+          "0x000000000000000000000000e66a957634bf7805a88dd50095d1b272cbaef155"
         ],
         "data": "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff",
-        "logIndex": 9,
-        "blockHash": "0x85fc97e390003ea4d36c5f9662709d4d81d5d4347460b20f494acd3450dd4d95"
+        "logIndex": 36,
+        "blockHash": "0xf5758601d9c942d6a26fb2a6647c7a1c192fc9ec9490d7f731ebde5f089c3506"
       }
     ],
-    "blockNumber": 12531895,
-    "cumulativeGasUsed": "6531515",
+    "blockNumber": 12571159,
+    "cumulativeGasUsed": "8778180",
     "status": 1,
     "byzantium": true
   },
   "args": [
-    "0x60E8aEAD0b8e149633B240eA948Ed50a81696750",
+    "0xA44567Ea4b9649c86D18ae7EdCB03d1752e792f7",
     "0x8f592F2594C43eA30879f9fbED4d930248D81D41",
-    "0x3520A7E6D178b8F41BBe78766C6A939434bacF69",
-    "0x99Fda92878c1d2f1e0971D1937C50CC578A33E3D"
+    "0xe66A957634bf7805a88dD50095d1b272cbAeF155",
+    "0x8D6C54d308C043Ee1914a4604d873E6538713AEF"
   ],
   "numDeployments": 1,
   "solcInputHash": "d84a26b853467fe455f16341e44935fd",
