+++ deployments/ropsten/YearnVault.json
@@ -1,5 +1,5 @@
 {
-  "address": "0x3520A7E6D178b8F41BBe78766C6A939434bacF69",
+  "address": "0xe66A957634bf7805a88dD50095d1b272cbAeF155",
   "abi": [
     {
       "inputs": [
@@ -372,26 +372,26 @@
       "type": "function"
     }
   ],
-  "transactionHash": "0xe5f4ffd1a755c07995e714f070ce995fa899cab70d865dc2883b96d5b941b7f1",
+  "transactionHash": "0x64609efcce008e2d0ca470e3cad66e8f0ef16bfad15c7e50926148b1b77c3574",
   "receipt": {
     "to": null,
     "from": "0x8f592F2594C43eA30879f9fbED4d930248D81D41",
-    "contractAddress": "0x3520A7E6D178b8F41BBe78766C6A939434bacF69",
-    "transactionIndex": 8,
+    "contractAddress": "0xe66A957634bf7805a88dD50095d1b272cbAeF155",
+    "transactionIndex": 5,
     "gasUsed": "1148247",
     "logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-    "blockHash": "0x20357cafa145fda0fc5544d68c6b5650feedb2f9da8da52983b390cd61f9e197",
-    "transactionHash": "0xe5f4ffd1a755c07995e714f070ce995fa899cab70d865dc2883b96d5b941b7f1",
+    "blockHash": "0x098f8ffdf04bbfdb2c5f5faf39d9ac15e03ad0cca0858de7f83a9116ec8ee31e",
+    "transactionHash": "0x64609efcce008e2d0ca470e3cad66e8f0ef16bfad15c7e50926148b1b77c3574",
     "logs": [],
-    "blockNumber": 12531894,
-    "cumulativeGasUsed": "3270575",
+    "blockNumber": 12571158,
+    "cumulativeGasUsed": "1774497",
     "status": 1,
     "byzantium": true
   },
   "args": [
     "LUSD Yearn Vault",
     "LUSD",
-    "0x99Fda92878c1d2f1e0971D1937C50CC578A33E3D"
+    "0x8D6C54d308C043Ee1914a4604d873E6538713AEF"
   ],
   "numDeployments": 1,
   "solcInputHash": "d84a26b853467fe455f16341e44935fd",
