--- deployments/ropsten/aUST.json
@@ -1,1 +1,0 @@
-{ "address": "0x006479f75d6622ae6a21be17c7f555b94c672342" }
